Now able to detect multiple retroreflective bands and correlate data.
The new algorithm check for aligned rectangles. If a group of rectangles have similar rotation values, they will be integrated in a structre. In order to improve robustness, the center calculation is done by averaging pixel positions, assuming a perfect rectangle.

This algorithm can also account for partially occuluded strips